test(hooks): add hooks to universalize tests
