+++ generator/test/basic/basic_test.go
@@ -4,12 +4,12 @@ package basic
 
 import (
 	"context"
-	"fmt"
 	"log"
-	"os/exec"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+
+	"github.com/prisma/photongo/generator/test/hooks"
 )
 
 type cx = context.Context
@@ -17,23 +17,6 @@ type Func func(t *testing.T, client Client, ctx cx)
 
 func str(v string) *string {
 	return &v
-}
-
-func cmd(name string, args ...string) error {
-	cmd := exec.Command(name, args...)
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		exit, ok := err.(*exec.ExitError)
-		if !ok {
-			return fmt.Errorf("command %s %s failed: %w", name, args, err)
-		}
-
-		if !exit.Success() {
-			return fmt.Errorf("%s %s exited with status code %d and output %s: %w", name, args, exit.ExitCode(), string(out), err)
-		}
-	}
-
-	return nil
 }
 
 func TestBasic(t *testing.T) {
@@ -273,19 +256,7 @@ func TestBasic(t *testing.T) {
 	for _, tt := range tests {
 		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
-			if err := cmd("rm", "-rf", "dev.sqlite"); err != nil {
-				log.Fatal(err)
-			}
-			if err := cmd("rm", "-rf", "migrations"); err != nil {
-				log.Fatal(err)
-			}
-
-			if err := cmd("prisma2", "lift", "save", "--create-db", "--name", "init"); err != nil {
-				t.Fatalf("could not run lift save %s", err)
-			}
-			if err := cmd("prisma2", "lift", "up"); err != nil {
-				t.Fatalf("could not run lift up %s", err)
-			}
+			hooks.Run(t)
 
 			client := NewClient()
 			if err := client.Connect(); err != nil {
